requirements.txt
Django~=3.1.4
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
Pillow~=8.1.0
django-heroku==0.3.1
gunicorn==20.0.4
django-summernote
django-cloudinary-storage
dj-database-url
dj-static
python-decouple
django-environ
gunicorn
psycopg2-binary
whitenoise